+++ cli/azd/pkg/infra/provisioning/manager_test.go
@@ -10,6 +10,8 @@ import (
 	"testing"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/policy"
+	"github.com/azure/azure-dev/cli/azd/internal/tracing"
+	"github.com/azure/azure-dev/cli/azd/internal/tracing/fields"
 	"github.com/azure/azure-dev/cli/azd/pkg/account"
 	"github.com/azure/azure-dev/cli/azd/pkg/azapi"
 	"github.com/azure/azure-dev/cli/azd/pkg/cloud"
@@ -394,4 +396,141 @@ func registerContainerDependencies(mockContext *mocks.MockContext, env *environm
 
 func defaultProvider() (provisioning.ProviderKind, error) {
 	return provisioning.Bicep, nil
+}
+
+func TestRecordInfraProviderUsage(t *testing.T) {
+	failingResolver := func() (provisioning.ProviderKind, error) {
+		return provisioning.NotSpecified, errors.New("no default provider")
+	}
+	unspecifiedResolver := func() (provisioning.ProviderKind, error) {
+		return provisioning.NotSpecified, nil
+	}
+
+	tests := []struct {
+		name            string
+		layers          []provisioning.Options
+		defaultProvider provisioning.DefaultProviderResolver
+		expected        string // "" means no infra.provider attribute is recorded
+	}{
+		{
+			name:            "single explicit bicep",
+			layers:          []provisioning.Options{{Provider: provisioning.Bicep}},
+			defaultProvider: defaultProvider,
+			expected:        "bicep",
+		},
+		{
+			name:            "single explicit terraform",
+			layers:          []provisioning.Options{{Provider: provisioning.Terraform}},
+			defaultProvider: defaultProvider,
+			expected:        "terraform",
+		},
+		{
+			name:            "unspecified resolves through default",
+			layers:          []provisioning.Options{{Provider: provisioning.NotSpecified}},
+			defaultProvider: defaultProvider,
+			expected:        "bicep",
+		},
+		{
+			name: "uniform provider across layers",
+			layers: []provisioning.Options{
+				{Provider: provisioning.Bicep},
+				{Provider: provisioning.Bicep},
+			},
+			defaultProvider: defaultProvider,
+			expected:        "bicep",
+		},
+		{
+			name: "different providers across layers records mixed",
+			layers: []provisioning.Options{
+				{Provider: provisioning.Bicep},
+				{Provider: provisioning.Terraform},
+			},
+			defaultProvider: defaultProvider,
+			expected:        provisioning.InfraProviderMixed,
+		},
+		{
+			name:            "single explicit arm",
+			layers:          []provisioning.Options{{Provider: provisioning.Arm}},
+			defaultProvider: defaultProvider,
+			expected:        "arm",
+		},
+		{
+			name:            "custom provider is bucketed",
+			layers:          []provisioning.Options{{Provider: provisioning.ProviderKind("my-extension-provider")}},
+			defaultProvider: defaultProvider,
+			expected:        provisioning.InfraProviderCustom,
+		},
+		{
+			name: "built-in plus custom records mixed",
+			layers: []provisioning.Options{
+				{Provider: provisioning.Bicep},
+				{Provider: provisioning.ProviderKind("my-extension-provider")},
+			},
+			defaultProvider: defaultProvider,
+			expected:        provisioning.InfraProviderMixed,
+		},
+		{
+			name: "two distinct custom providers record mixed",
+			layers: []provisioning.Options{
+				{Provider: provisioning.ProviderKind("vendor.one")},
+				{Provider: provisioning.ProviderKind("vendor.two")},
+			},
+			defaultProvider: defaultProvider,
+			expected:        provisioning.InfraProviderMixed,
+		},
+		{
+			name:            "no layers records nothing",
+			layers:          nil,
+			defaultProvider: defaultProvider,
+			expected:        "",
+		},
+		{
+			name:            "unspecified with nil resolver records nothing",
+			layers:          []provisioning.Options{{Provider: provisioning.NotSpecified}},
+			defaultProvider: nil,
+			expected:        "",
+		},
+		{
+			name:            "unspecified with failing resolver records nothing",
+			layers:          []provisioning.Options{{Provider: provisioning.NotSpecified}},
+			defaultProvider: failingResolver,
+			expected:        "",
+		},
+		{
+			name:            "unspecified resolving to NotSpecified records nothing",
+			layers:          []provisioning.Options{{Provider: provisioning.NotSpecified}},
+			defaultProvider: unspecifiedResolver,
+			expected:        "",
+		},
+	}
+
+	// Usage attributes are process-global, so these subtests must not run in parallel.
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			tracing.ResetUsageAttributesForTest()
+			t.Cleanup(tracing.ResetUsageAttributesForTest)
+
+			// RecordInfraProviderUsage only reads the manager's default provider, so the remaining
+			// dependencies are intentionally nil.
+			mgr := provisioning.NewManager(nil, tt.defaultProvider, nil, nil, nil, nil, nil, nil)
+			mgr.RecordInfraProviderUsage(tt.layers)
+
+			var got string
+			var found bool
+			for _, attr := range tracing.GetUsageAttributes() {
+				if attr.Key == fields.InfraProviderKey.Key {
+					got = attr.Value.AsString()
+					found = true
+				}
+			}
+
+			if tt.expected == "" {
+				require.False(t, found, "expected no infra.provider attribute, got %q", got)
+				return
+			}
+
+			require.True(t, found, "expected infra.provider attribute to be recorded")
+			require.Equal(t, tt.expected, got)
+		})
+	}
 }
